# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "aziz16110904@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/branches": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Get all branches",
                "description": "Retrieve a list of all branches with associated images.",
                "operationId": "8bad73dccf3e88f9ee9007d8d3eb1040",
                "responses": {
                    "200": {
                        "description": "List of branches",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Branch Name"
                                                    },
                                                    "brand": {
                                                        "type": "string",
                                                        "example": "Tech"
                                                    },
                                                    "region": {
                                                        "type": "string",
                                                        "example": "Tashkent"
                                                    },
                                                    "district": {
                                                        "type": "string",
                                                        "example": "Toshkent"
                                                    },
                                                    "images": {
                                                        "type": "array",
                                                        "items": {
                                                            "description": "URLs of associated images",
                                                            "type": "string",
                                                            "format": "url"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Branches"
                ],
                "summary": "Create a new branch with optional image uploads",
                "description": "Creates a new branch with specified details and optional images.",
                "operationId": "df88a7dda440ba67d2d3433731a1dbee",
                "requestBody": {
                    "description": "Branch data with optional image upload (multi-select)",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "brand_id",
                                    "region_id",
                                    "district_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Branch Name"
                                    },
                                    "brand_id": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": "1"
                                    },
                                    "region_id": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": "1"
                                    },
                                    "district_id": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": "15"
                                    },
                                    "images[]": {
                                        "description": "Array of image files (JPEG, PNG, GIF) to upload",
                                        "type": "array",
                                        "items": {
                                            "description": "Image file (JPEG, PNG, GIF)",
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Branch created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Branch created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Branch Name"
                                                },
                                                "brand": {
                                                    "type": "string",
                                                    "example": "Tech"
                                                },
                                                "region": {
                                                    "type": "string",
                                                    "example": "Tashkent"
                                                },
                                                "district": {
                                                    "type": "string",
                                                    "example": "Toshkent"
                                                },
                                                "images": {
                                                    "type": "array",
                                                    "items": {
                                                        "description": "URLs of uploaded images",
                                                        "type": "urls",
                                                        "format": "url"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}